Create DailyFolder
Set up a task scheduler: Open the Task Scheduler by searching for it in the Start menu, and click on "Create Basic Task" from the right-hand pane. Follow the prompts to create a task that will run the batch file at the desired time of the day.
Create a batch file: Open Notepad and create a new text file. In the file, type the commands from steps 2-4. Save the file with a .bat extension. For example, you can save the file as "CreateDailyFolder.bat".
Add the batch file to the task scheduler: In the task scheduler, click on "Action" from the right-hand pane and select "Create Task". In the "General" tab, give your task a name and select the user account that the task will run under. In the "Triggers" tab, set up the schedule for the task. In the "Actions" tab, click on "New" and select "Start a program". In the "Program/script" field, browse to the batch file you created in step 6 and click on "OK".